13.7.0
- themekit support, create themekits with powertags,weaknesstags,quest and special improvements. Easy import of tags via a fast CSV import
- added themekit menu item to the actor sheet (top right 3 point menu)
- added themekit assignment for empty characters
- assign & remove themekits to exisiting themebooks
- add powertags from themekits to themebooks
- faster deletion of tags & statuses

+++ module/sheets/actor-sheet.mjs
@@ -287,8 +287,26 @@ export class MistEngineActorSheet extends HandlebarsApplicationMixin(ActorSheetV
     static async #handleDeleteFloatingTagOrStatus(event, target) {
         event.preventDefault();
         const index = target.dataset.index;
-        await FloatingTagAndStatusAdapter.handleDeleteFloatingTagOrStatus(this.actor, index);
-        this.sendFloatableTagOrStatusUpdate();
+         let confirmed = true;
+        if(target.dataset.confirm && target.dataset.confirm == "1"){
+            confirmed = await foundry.applications.api.DialogV2.confirm({
+                title: "Confirm Deletion",
+                content: "Are you sure you want to delete this tag/status?",
+                yes: {
+                    label: "Yes",
+                    callback: () => true
+                },
+                no: {
+                    label: "No",
+                    callback: () => false
+                }
+            });   
+        }
+        
+        if(confirmed){
+            await FloatingTagAndStatusAdapter.handleDeleteFloatingTagOrStatus(this.actor, index);
+            this.sendFloatableTagOrStatusUpdate();
+        }
     }
 
     async handleFtStatChanged(event) {
